requirements.txt
streamlit
langchain
langgraph
langchain-nvidia-ai-endpoints
faiss-cpu
nemoguardrails
mcp
fastmcp
python-dotenv
beautifulsoup4
pypdf
fastapi
uvicorn
pydantic
starlette
